move field_importer.py from core to io
