+++ exetera/core/operations.py
@@ -2613,7 +2613,7 @@ def get_byte_map(string_map):
 @njit           
 def categorical_transform(chunk, i_c, column_inds, column_vals, column_offsets, cat_keys, cat_index, cat_values):
     """
-    Tranform method for categorical importer in readerwriter.py
+    Transform method for categorical importer in readerwriter.py
     """   
     col_offset = column_offsets[i_c]
 
@@ -2644,7 +2644,7 @@ def categorical_transform(chunk, i_c, column_inds, column_vals, column_offsets, 
 @njit           
 def leaky_categorical_transform(chunk, freetext_indices, freetext_values, i_c, column_inds, column_vals, column_offsets, cat_keys, cat_index, cat_values):
     """
-    Tranform method for categorical importer in readerwriter.py
+    Transform method for categorical importer in readerwriter.py
     """   
     col_offset = column_offsets[i_c] 
 
@@ -2794,7 +2794,7 @@ def raiseNumericException(exception_message, exception_args):
 def transform_int(column_inds, column_vals, column_offsets, col_idx,
                     written_row_count, invalid_value, validation_mode, data_type, field_name):
     """
-    Tranform int method for numeric importer in field_importer.py
+    Transform int method for numeric importer in field_importer.py
     """
     widths = column_inds[col_idx, 1:written_row_count + 1] - column_inds[col_idx, :written_row_count]
     width = widths.max()
@@ -2839,7 +2839,7 @@ def transform_int(column_inds, column_vals, column_offsets, col_idx,
 def transform_float(column_inds, column_vals, column_offsets, col_idx,
                       written_row_count, invalid_value, validation_mode, data_type, field_name):
     """
-    Tranform float method for numeric importer in field_importer.py
+    Transform float method for numeric importer in field_importer.py
     """
     widths = column_inds[col_idx, 1:written_row_count + 1] - column_inds[col_idx, :written_row_count]
     width = widths.max()
@@ -2899,7 +2899,7 @@ def transform_to_values(column_inds, column_vals, column_offsets, col_idx, writt
 def fixed_string_transform(column_inds, column_vals, column_offsets, col_idx, written_row_count,
                            strlen, memory):
     """
-    Tranform method for fixed string importer in field_importer.py
+    Transform method for fixed string importer in field_importer.py
     """
     col_offset = column_offsets[col_idx]
     for i in range(written_row_count):
