+++ tests/test_fields.py
@@ -9,7 +9,7 @@ from datetime import datetime
 from exetera.core import session
 from exetera.core import fields
 from exetera.core import persistence as per
-from exetera.core import field_importers as fi
+from exetera.io import field_importers as fi
 from exetera.core import utils
